chore: arrange utils better
