+++ src/components/--MetaGroup/MetaGroup.js
@@ -3,21 +3,13 @@ import PropTypes from "prop-types";
 
 import { Box } from "grommet";
 
+import { childrenWithProps } from "../../common/reactUtils";
+
 /**
  * A wrapper for meta tags, providing separators.
  * To be placed in post header.
  */
 const MetaGroup = ({ children }) => {
-  const childrenWithProps = React.Children.map(children, (child) => {
-    if (React.isValidElement(child)) {
-      return React.cloneElement(child, {
-        color: "text-invert",
-        size: "small",
-        weight: "bold",
-      });
-    }
-    return child;
-  });
   return (
     <>
       <Box
@@ -29,7 +21,11 @@ const MetaGroup = ({ children }) => {
         color="text-invert"
         alignContent="center"
       >
-        {childrenWithProps}
+        {childrenWithProps(children, {
+          color: "text-invert",
+          size: "small",
+          weight: "bold",
+        })}
       </Box>
     </>
   );
