+++ src/components/--WidgetContainer/WidgetContainer.js
@@ -3,7 +3,8 @@ import PropTypes from "prop-types";
 
 import { Box, ResponsiveContext } from "grommet";
 
-import { childrenWithProps, isObject } from "../../utils";
+import { childrenWithProps } from "../../common/reactUtils";
+import { isObject } from "../../common/utils";
 
 /**
  * A responsive container to hold post previews, based on grommet <Box>
